+++ elmyra.ip.access.epo/elmyra/ip/access/epo/static/js/app/components/basket.js
@@ -5,48 +5,105 @@ BasketModel = Backbone.RelationalModel.extend({
 
     sync: Backbone.localforage.sync('Basket'),
 
+    relations: [
+        {
+            type: Backbone.HasMany,
+            key: 'entries',
+            relatedModel: 'BasketEntryModel',
+            autoFetch: true,
+            includeInJSON: Backbone.Model.prototype.idAttribute,
+
+            /*
+            reverseRelation: {
+                type: Backbone.One,
+                key: 'basket',
+                // 'relatedModel' is automatically set to 'ProjectModel'
+                includeInJSON: Backbone.Model.prototype.idAttribute,
+            },
+            */
+
+        }
+    ],
+
     defaults: {
-        numberlist: [],
     },
 
     initialize: function() {
         console.log('BasketModel.initialize');
-        this.init_from_query();
     },
 
     // initialize model from url query parameters
     init_from_query: function() {
+        var _this = this;
         var url = $.url(window.location.href);
         var attribute_name = 'numberlist';
         var value = url.param(attribute_name);
         if (value) {
             value = decodeURIComponent(value);
-            value = value.split(',');
-            this.set(attribute_name, value);
+            var entries = value.split(/[,\n]/);
+            _(entries).each(function(entry) {
+                _this.add(entry.trim());
+            });
+        }
+    },
+
+    get_entry_by_number: function(item) {
+        var entrymodels = this.get('entries').where({number: item});
+        if (_.isEmpty(entrymodels)) {
+            return;
+        } else {
+            return entrymodels[0];
         }
     },
 
     // add item to basket
     add: function(item) {
-        var numberlist = this.get('numberlist');
-        if (_(numberlist).contains(item)) { return; }
-        numberlist.push(item);
-        this.trigger('change', this);
-        this.trigger('change:add', item);
+        var _this = this;
+
+        if (this.get_entry_by_number(item)) {
+            return;
+        }
+
+        var entries = this.get('entries');
+        var entry = new BasketEntryModel({number: item /*, basket: this*/});
+        entry.save(null, {success: function() {
+            entries.add(entry);
+            _this.save({'entries': entries}, {success: function() {
+                $.when(_this.fetch_entries()).then(function() {
+                    _this.trigger('change', _this);
+                    _this.trigger('change:add', item);
+                });
+            }});
+        }});
     },
 
     // remove item from basket
     remove: function(item) {
-        var numberlist = this.get('numberlist');
-        if (!_(numberlist).contains(item)) { return; }
-        this.set('numberlist', _.without(numberlist, item));
-        //this.trigger('change', this);
-        this.trigger('change:remove', item);
+        var _this = this;
+
+        var entry = this.get_entry_by_number(item);
+        if (!entry) {
+            return;
+        }
+
+        var entries = this.get('entries');
+        entries.remove(entry);
+        entry.destroy();
+        _this.save({'entries': entries}, {success: function() {
+            $.when(_this.fetch_entries()).then(function() {
+                _this.trigger('change:remove', item);
+                _this.trigger('change', _this);
+            });
+        }});
+    },
+
+    get_numbers: function() {
+        return this.get('entries').invoke('get', 'number');
     },
 
     review: function(options) {
 
-        // compute cql query from numberlist in basket
+        // compute cql query from entries in basket
         var basket = $('#basket').val();
         if (!basket) {
             return;
@@ -63,8 +120,72 @@ BasketModel = Backbone.RelationalModel.extend({
         opsChooserApp.set_datasource('review');
         opsChooserApp.metadata.set('reviewmode', true);
         opsChooserApp.perform_listsearch(options, query, publication_numbers, hits, 'pn', 'OR');
-    }
+    },
 
+    // fetch all basket entries from datastore, one by one; this is nasty
+    fetch_entries: function() {
+
+        var _this = this;
+        var main_deferred = $.Deferred();
+        $.when(this.fetchRelated('entries')).then(function() {
+
+            var deferreds = [];
+            _this.get('entries').each(function(entry) {
+
+                // prepare a deferred which will get resolved after successfully fetching an item from datastore
+                var deferred = $.Deferred();
+                deferreds.push(deferred.promise());
+
+                entry.fetch({
+                    success: function() {
+                        deferred.resolve(entry);
+                    },
+                    error: function() {
+                        // HACK: sometimes, the item has vanished while fetching from store, so let's recreate it
+                        console.log('error while fetching basket entry:', entry);
+                        entry.save(null, {
+                            success: function() {
+                                console.log('success');
+                                deferred.resolve(entry);
+                            },
+                            error: function() {
+                                console.log('error');
+                                deferred.resolve(entry);
+                            },
+                        });
+                    }
+                });
+            });
+
+            $.when.apply($, deferreds).then(function() {
+                main_deferred.resolve();
+            });
+        });
+
+        return main_deferred.promise();
+
+    },
+
+});
+
+
+BasketEntryModel = Backbone.RelationalModel.extend({
+
+    sync: Backbone.localforage.sync('BasketEntry'),
+
+    defaults: {
+        number: undefined,
+        timestamp: undefined,
+        title: undefined,
+        rating: undefined,
+        dismiss: undefined,
+        // TODO: link to QueryModel
+        //query: undefined,
+    },
+
+    initialize: function() {
+        console.log('BasketEntryModel.initialize');
+    },
 });
 
 BasketView = Backbone.Marionette.ItemView.extend({
@@ -78,20 +199,25 @@ BasketView = Backbone.Marionette.ItemView.extend({
     },
 
     serializeData: function() {
+
+        var _this = this;
+
         var data = {};
         data = this.model.toJSON();
 
-        var numberlist = this.model.get('numberlist');
-        if (numberlist) {
-            data['numberlist'] = numberlist.join('\n');
-        }
-
         _(data).extend(this.params_from_query());
 
+        var numbers = this.model.get_numbers();
+        if (numbers) {
+            data['numbers_display'] = numbers.join('\n');
+        }
+
         return data;
+
     },
 
     // initialize more template variables from url query parameters
+    // TODO: refactor to utils.js modulo basket_option_names
     params_from_query: function() {
         var tplvars = {};
         var basket_option_names = ['ship-url', 'ship-param'];
@@ -145,16 +271,16 @@ BasketView = Backbone.Marionette.ItemView.extend({
 
             var projectname = opsChooserApp.project.get('name');
 
-            var numberlist = _this.model.get('numberlist');
-            var numberlist_count = numberlist.length;
-            var numberlist_string = numberlist.join('\n');
+            var numbers = _this.model.get_numbers();
+            var numbers_count = numbers.length;
+            var numbers_string = numbers.join('\n');
 
             var subject = _.template('[IPSUITE] Shared <%= count %> patent numbers through project <%= projectname %> at <%= date %>')({
-                count: numberlist_count,
+                count: numbers_count,
                 date: now_iso_human(),
                 projectname: projectname,
             });
-            var body = numberlist_string + '\r\n\r\n--\r\nPowered by https://patentsearch.elmyra.de/';
+            var body = numbers_string + '\r\n\r\n--\r\nPowered by https://patentsearch.elmyra.de/';
             var mailto_link = _.template('mailto:?subject=<%= subject %>&body=<%= body %>')({
                 subject: encodeURIComponent(subject),
                 body: encodeURIComponent(body),
@@ -166,7 +292,7 @@ BasketView = Backbone.Marionette.ItemView.extend({
         });
 
         // display number of entries in basket
-        var entry_count = this.model.get('numberlist').length;
+        var entry_count = this.model.get('entries').length;
         $('.basket-entry-count').text(entry_count);
 
     },
@@ -183,6 +309,33 @@ BasketView = Backbone.Marionette.ItemView.extend({
 
     onDomRefresh: function() {
         console.log('BasketView.onDomRefresh');
+    },
+
+    // backpropagate current basket entries into checkbox state
+    link_document: function(entry) {
+
+        // why do we have to access the global object here?
+        // maybe because of the event machinery which dispatches to us?
+        var numbers = opsChooserApp.basketModel.get_numbers();
+
+        var checkbox_element = $('#' + 'chk-patent-number-' + entry);
+        var add_button_element = $('#' + 'add-patent-number-' + entry);
+        var remove_button_element = $('#' + 'remove-patent-number-' + entry);
+
+        // number is not in basket, show "add" button
+        if (!_(numbers).contains(entry)) {
+            checkbox_element && checkbox_element.prop('checked', false);
+            add_button_element && add_button_element.show();
+            remove_button_element && remove_button_element.hide();
+
+        // number is already in basket, show "remove" button
+        } else {
+            checkbox_element && checkbox_element.prop('checked', true);
+            add_button_element && add_button_element.hide();
+            remove_button_element && remove_button_element.show();
+
+        }
+
     },
 
 });