+++ elmyra.ip.access.epo/elmyra/ip/access/epo/static/js/app/views/ops.js
@@ -38,7 +38,7 @@ OpsExchangeDocumentView = Backbone.Marionette.ItemView.extend({
     // e.g. to bind click handlers on individual records
     onDomRefresh: function() {
         var patent_number = this.model.attributes.get_patent_number();
-        opsChooserApp.collectionView.basket_update_ui_entry(patent_number);
+        opsChooserApp.basketView.link_document(patent_number);
     },
 
 });
@@ -52,26 +52,6 @@ OpsExchangeDocumentCollectionView = Backbone.Marionette.CompositeView.extend({
 
     appendHtml: function(collectionView, itemView) {
         $(collectionView.el).append(itemView.el);
-    },
-
-    // backpropagate current basket entries into checkbox state
-    // TODO: maybe refactor this elsewhere
-    basket_update_ui_entry: function(entry) {
-        // TODO: move to model access
-        //console.log(this.model);
-        var payload = $('#basket').val();
-        var checkbox_element = $('#' + 'chk-patent-number-' + entry);
-        var add_button_element = $('#' + 'add-patent-number-' + entry);
-        var remove_button_element = $('#' + 'remove-patent-number-' + entry);
-        if (_.string.include(payload, entry)) {
-            checkbox_element && checkbox_element.prop('checked', true);
-            add_button_element && add_button_element.hide();
-            remove_button_element && remove_button_element.show();
-        } else {
-            checkbox_element && checkbox_element.prop('checked', false);
-            add_button_element && add_button_element.show();
-            remove_button_element && remove_button_element.hide();
-        }
     },
 
 });